+++ core/CHANGELOG.md
@@ -9,6 +9,8 @@
 - For code simplicity, also removed the ability to stop getting the context & the explanation. Now all code will look the same, always getting `result.response` instead of sometimes `result.response` and others just `result` (because of the return type being just the number and not the CLResponse object).
 - Now `source` was renamed to `sources` and is a `string[]`. `getSource` was renamed to `getSources` too.
 - Added the activeness factor to BMR calculation.
+- (hopefully) fixed accuracy of supported `performance/` calculations.
+- (WIP) update BMI percentiles from 2000 to 2022 data from the CDC.
 
 ## 0.1.0